hange to migrate syllabus to the admissions module and replace certificate with syllabus on the cohort model
