+++ breathecode/admissions/tests/urls/tests_certificate.py
@@ -53,6 +53,8 @@ class CertificateTestSuite(AdmissionsTestCase):
             'id': self.certificate.id,
             'name': self.certificate.name,
             'slug': self.certificate.slug,
+            'logo': self.certificate.logo,
+            'description': self.certificate.description,
         }])
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(self.count_certificate(), 1)
