+++ breathecode/admissions/urls.py
@@ -2,8 +2,9 @@ from django.contrib import admin
 from django.urls import path, include
 from rest_framework import routers
 from .views import (
-    AcademyView, CohortUserView, CertificateView, get_cohorts, AcademyCohortView,
-    get_timezones, UserView, UserMeView, AcademyCohortUserView
+    AcademyView, CohortUserView, get_cohorts, AcademyCohortView,
+    get_timezones, UserView, UserMeView, AcademyCohortUserView,
+    get_courses, get_single_course, SyllabusView
 )
 
 app_name = 'admissions'
@@ -27,7 +28,12 @@ urlpatterns = [
     path('user', UserView.as_view(), name="user"),
 
     # update a cohort user information
-    path('certificate/', CertificateView.as_view(), name="certificate"),
+    path('certificate', get_courses, name="certificate"),
+    path('certificate/<str:certificate_slug>/', get_single_course),
+    path('certificate/<str:certificate_slug>/syllabus', SyllabusView.as_view()),
+    path('certificate/<str:certificate_slug>/syllabus/<int:version>', SyllabusView.as_view()),
+    path('certificate/<str:certificate_slug>/academy/<int:academy_id>/syllabus/<int:version>', SyllabusView.as_view()),
+    path('certificate/<str:certificate_slug>/academy/<int:academy_id>/syllabus', SyllabusView.as_view()),
     
     path('catalog/timezones', get_timezones, name="timezones_all"),
 ]
