+++ breathecode/admissions/views.py
@@ -11,7 +11,8 @@ from .serializers import (
     AcademySerializer, CohortSerializer, CertificateSerializer,
     GetCohortSerializer, UserSerializer, CohortUserSerializer,
     GETCohortUserSerializer, CohortUserPUTSerializer, CohortPUTSerializer,
-    CohortUserPOSTSerializer, UserDJangoRestSerializer, UserMeSerializer
+    CohortUserPOSTSerializer, UserDJangoRestSerializer, UserMeSerializer,
+    GetCertificateSerializer
 )
 from .models import Academy, City, CohortUser, Certificate, Cohort, Country, STUDENT, DELETED
 from breathecode.authenticate.models import ProfileAcademy
@@ -140,7 +141,10 @@ class CohortUserView(APIView):
         return Response(serializer.data)
 
     def count_certificates_by_cohort(self, cohort, user_id):
-        return (CohortUser.objects.filter(user_id=user_id, cohort__certificate=cohort.certificate)
+        if cohort.syllabus is None:
+            return 0
+
+        return (CohortUser.objects.filter(user_id=user_id, cohort__syllabus__certificate=cohort.syllabus.certificate)
             .exclude(educational_status='POSTPONED').count())
 
     def validations(self, request, cohort_id=None, user_id=None, matcher=None,
@@ -284,7 +288,7 @@ class AcademyCohortUserView(APIView):
         return Response(serializer.data)
 
     def count_certificates_by_cohort(self, cohort, user_id):
-        return (CohortUser.objects.filter(user_id=user_id, cohort__certificate=cohort.certificate)
+        return (CohortUser.objects.filter(user_id=user_id, cohort__syllabus__certificate=cohort.syllabus.certificate)
             .exclude(educational_status='POSTPONED').count())
 
     def validations(self, request, cohort_id=None, user_id=None, matcher=None,
@@ -439,13 +443,13 @@ class AcademyCohortView(APIView):
         if academy is None:
             raise ValidationError(f'Academy {academy_id} not found')
 
-        certificate_id = request.data.get('certificate')
-        if certificate_id is None:
-            raise ParseError(detail='certificate field is missing')
+        syllabus_id = request.data.get('syllabus')
+        if syllabus_id is None:
+            raise ParseError(detail='syllabus field is missing')
 
-        certificate = Certificate.objects.filter(id=certificate_id).first()
-        if certificate is None:
-            raise ParseError(detail='specified certificate not be found')
+        syllabus = Syllabus.objects.filter(id=syllabus_id).first()
+        if syllabus is None:
+            raise ParseError(detail='specified syllabus not be found')
 
         if request.data.get('current_day'):
             raise ParseError(detail='current_day field is not allowed')
@@ -458,7 +462,7 @@ class AcademyCohortView(APIView):
         for key in request.data:
             data[key] = request.data.get(key)
 
-        data['certificate'] = certificate
+        data['syllabus'] = syllabus
 
         serializer = CohortSerializer(data=data, context=data)
         if serializer.is_valid():
@@ -516,3 +520,87 @@ class CertificateView(APIView):
         items = Certificate.objects.all()
         serializer = CertificateSerializer(items, many=True)
         return Response(serializer.data)
+
+@api_view(['GET'])
+def get_courses(request):
+    certificates = Certificate.objects.all()
+    serializer = GetCertificateSerializer(certificates, many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+@api_view(['GET'])
+def get_single_course(request, certificate_slug):
+    certificates = Certificate.objects.filter(slug=certificate_slug).first()
+    if certificates is None:
+        raise serializers.ValidationError("Certificate slug not found", code=404)
+    serializer = GetCertificateSerializer(certificates, many=False)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+class SyllabusView(APIView):
+    """
+    List all snippets, or create a new snippet.
+    """
+
+    # TODO: uncomment decorator when finally all students have a ProfileAcademy object.
+    # @capable_of('crud_syllabus')
+    def get(self, request, certificate_slug=None, version=None, academy_id=None):
+        if academy_id is None:
+            raise ValidationException("Missing academy id")
+        
+        certificates = Certificate.objects.filter(slug=certificate_slug).first()
+        if certificates is None:
+            raise serializers.ValidationError("Certificate slug not found", code=404)
+
+        syl = None
+        if version is None:
+            # only public syllabus or for the academy owner
+            syl = certificates.syllabus_set.filter(Q(academy_owner__id=academy_id) | Q(private=False))
+            serializer = SyllabusSmallSerializer(syl, many=True)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            # only public syllabus or for the academy owner
+            syl = certificates.syllabus_set.filter(version=version).filter(Q(academy_owner__id=academy_id) | Q(private=False)).first()
+        
+        if syl is None:
+            raise ValidationException("Syllabus not found", code=404)
+
+        # TODO: if a syllabus is private, no other academy should be able to see it
+
+        serializer = SyllabusGetSerializer(syl, many=False)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @capable_of('crud_syllabus')
+    def post(self, request, certificate_slug=None, academy_id=None):
+        version = 1
+        certificates = Certificate.objects.filter(slug=certificate_slug).first()
+        if certificates is None:
+            raise serializers.ValidationError(f"Invalid certificates slug {course__slug}", code=404)
+
+        item = Syllabus.objects.filter(course__slug=certificate_slug, academy_owner__id=academy_id).order_by('version').first()
+        if item is not None:
+            version = item.version + 1
+
+        academy = Academy.objects.filter(id=academy_id).first()
+        if academy is None:
+            raise ValidationException(f"Invalid academy {str(academy_id)}")
+
+        serializer = SyllabusSerializer(data=request.data, context={"certificates": certificates, "academy": academy })
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    @capable_of('crud_syllabus')
+    def put(self, request, certificate_slug=None, version=None, academy_id=None):
+        if version is None:
+            raise serializers.ValidationError("Missing syllabus version", code=400)
+        
+        item = Syllabus.objects.filter(course__slug=certificate_slug, version=version, academy_owner__id=academy_id).first()
+        if item is None:
+            raise serializers.ValidationError("Syllabus version not found for this academy", code=404)
+        
+        serializer = SyllabusSerializer(item, data=request.data, many=False)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
