+++ breathecode/urls.py
@@ -46,7 +46,6 @@ urlpatterns = [
     path('v1/events/', include('breathecode.events.urls', namespace='events')),
     path('v1/activity/', include('breathecode.activity.urls', namespace='activity')),
     path('v1/feedback/', include('breathecode.feedback.urls', namespace='feedback')),
-    path('v1/coursework/', include('breathecode.coursework.urls', namespace='coursework')),
     path('v1/messaging/', include('breathecode.notify.urls', namespace='notify')),
     path('v1/certificate/', include('breathecode.certificate.urls', namespace='certificate')),
